fix(ios): disambiguate encryption domain states
